fix handling of identity transformations
Unlike the assumption in 1fea7870fd5c8e7ccb8c4807fc31384702b24534,
an empty chain is *NOT* an invalid transformation object as it
is possible to have the same source and target frame (i.e. an identity
transformation). Whether a transformation object has a chain is tracked
by the 'valid' flag.